feat: create RowNavigation as a custom module
