+++ log-viewer/modules/calltree-view/CalltreeView.ts
@@ -10,6 +10,7 @@ import '../../resources/css/DatabaseView.scss';
 import '../../resources/css/TreeView.css';
 import { rootMethod } from '../Main';
 import { showTab } from '../Util';
+import { RowNavigation } from '../datagrid/module/RowNavigation';
 import { LogLine, RootNode, TimedNode } from '../parsers/TreeParser';
 import { hostService } from '../services/VSCodeService';
 
@@ -20,82 +21,7 @@ export async function renderCallTree(rootMethod: RootNode): Promise<void> {
     return Promise.resolve();
   }
 
-  type Module = { [key: string]: unknown };
-  const rowNavKeyBindings: Module = {
-    previousRow: function () {
-      const table = this.table as Tabulator;
-      const row = table.getSelectedRows()[0];
-      const previousRow = row?.getPrevRow();
-      if (previousRow) {
-        table.blockRedraw();
-        row.deselect();
-        previousRow.select();
-        table.restoreRedraw();
-        previousRow.getElement().scrollIntoView({ block: 'nearest' });
-      }
-    },
-    nextRow: function () {
-      const table = this.table as Tabulator;
-      const row = table.getSelectedRows()[0];
-      const nextRow = row?.getNextRow();
-      if (nextRow) {
-        table.blockRedraw();
-        row.deselect();
-        nextRow.select();
-        table.restoreRedraw();
-        nextRow.getElement().scrollIntoView({ block: 'nearest' });
-      }
-    },
-    expandRow: function () {
-      const table = this.table as Tabulator;
-      const row = table.getSelectedRows()[0];
-      if (!row) {
-        return;
-      }
-
-      if (row.isTreeExpanded()) {
-        const nextRow = row?.getNextRow();
-        if (nextRow && nextRow.getTreeParent() === row) {
-          table.blockRedraw();
-          row.deselect();
-          nextRow.select();
-          table.restoreRedraw();
-          nextRow.getElement().scrollIntoView({ block: 'nearest' });
-        }
-      } else {
-        row.treeExpand();
-      }
-    },
-    collapseRow: function () {
-      const table = this.table as Tabulator;
-      const row = table.getSelectedRows()[0];
-      if (!row) {
-        return;
-      }
-
-      if (!row.isTreeExpanded()) {
-        const prevRow = row?.getTreeParent();
-        if (prevRow) {
-          table.blockRedraw();
-          row.deselect();
-          prevRow.select();
-          table.restoreRedraw();
-          prevRow.getElement().scrollIntoView({ block: 'nearest' });
-        }
-      } else {
-        row.treeCollapse();
-      }
-    },
-  };
-
-  Tabulator.extendModule('keybindings', 'actions', rowNavKeyBindings);
-  Tabulator.extendModule('keybindings', 'bindings', {
-    previousRow: '38',
-    nextRow: '40',
-    expandRow: '39',
-    collapseRow: '37',
-  });
-
+  Tabulator.registerModule(RowNavigation);
   calltreeTable = new Tabulator('#calltreeTable', {
     data: toCallTree(rootMethod.children),
     layout: 'fitColumns',
@@ -272,10 +198,15 @@ export async function renderCallTree(rootMethod: RootNode): Promise<void> {
     }
   });
 
-  document.getElementById('calltreeTable')?.addEventListener(
+  //todo: move to my custom keybindings module
+  document.querySelector('#calltreeTable')?.addEventListener(
     'keydown',
     function (e) {
-      if (['ArrowDown', 'ArrowUp', 'ArrowLeft', 'ArrowRight', 'Space'].indexOf(e.key) > -1) {
+      const targetElem = e.target as HTMLElement;
+      if (
+        targetElem.classList.contains('tabulator-tableholder') &&
+        ['ArrowDown', 'ArrowUp', 'ArrowLeft', 'ArrowRight', 'Space'].indexOf(e.key) > -1
+      ) {
         e.preventDefault();
       }
     },
